add options that allow for more accurate index migration
* When migrating indexes, one will usually enable dual writes and then
begin a stream2es to stream old index data to the new index. In this
scenario, if a new write has written a document before stream2es attempts
to write it, we'd rather the write error and be skipped. This commit allows
us to use the `create` operation to avoid overwriting existing documents with the
`-no-clobber` flag.
* When using this option, many `error-*` documents would be created, allow disabling
the creation of these with the --no-tee-errors flag
* Fix an issue where error counts were not properly being tracked because the error-count
was being nil'd out before being added to the state
* Fix an issue where the trace log in `index-bulk` reported innacurate results